refactor: adicionando a verificação de autoridade para executar a requesição.
